# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CyCove Backend API
  version: 0.6.0-draft
  description: >
    HTTP API contract for the CyCove relay backend, formalizing the surface
    sketched in the vault note (Projects/CyCove.md, Architecture -> Backend).
    The server is a deliberate "dumb pipe": it never sees plaintext message
    content and never holds long-term encryption keys. Realtime delivery is
    covered separately in asyncapi.yaml.
    STATUS (2026-07-19): reshaped around matrix-sdk-crypto-wasm's actual
    protocol requirements — see docs/crypto-integration-notes.md for why.
    /keys/upload, /keys/query, /keys/claim replace the original /prekeys
    endpoints (Matrix-shaped batch operations, not this API's own simpler
    shapes). user_id and device_id are now client-generated, not
    server-assigned — required because the client's OlmMachine bakes both
    into a self-signed device_keys blob before /register is ever called.
    STATUS (2026-07-19, later same day): added a real username directory
    (/users/*), rate limiting on the most abuse-prone endpoints, and opaque
    encrypted contact backup (/account-data/*) — see
    docs/crypto-integration-notes.md and THREAT_MODEL.md for the disclosures
    that come with both.
    STATUS (2026-07-20): added /login (re-authenticate an existing account on
    the device that created it) and wired the web client to the previously
    backend-only /devices/pairing-token and /devices/link (real "linked
...
    more use for it.
...
servers:
...
paths:
...
  /keys/upload:
    post:
      tags: [keys]
      summary: Upload/replenish this device's keys
      description: >
        Matches matrix-sdk-crypto-wasm's KeysUploadRequest exactly — see
        docs/crypto-integration-notes.md. Used after registration mainly to
        top up the one-time-key pool as it's consumed; device_keys can be
        included too but rarely changes after the first upload. The device
        is taken from the bearer token, not the request body.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_keys:
                  type: object
                  description: Opaque — only needed if updating identity keys, which is rare after registration.
                one_time_keys:
                  type: object
                  description: Opaque map of "signed_curve25519:<keyId>" -> signed key object to add to the pool.
                fallback_keys:
                  type: object
      responses:
        '200':
          description: Keys stored
          content:
            application/json:
              schema:
                type: object
                required: [one_time_key_counts]
                properties:
                  one_time_key_counts:
                    type: object
                    description: >
                      Required by matrix-sdk-crypto-wasm's markRequestAsSent
                      (throws without this field) — confirmed by triggering
                      the error directly during implementation.
                    properties:
                      signed_curve25519:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /keys/query:
    post:
      tags: [keys]
      summary: Batch-fetch device_keys for a set of users
      description: Matches matrix-sdk-crypto-wasm's KeysQueryRequest exactly — issued even for the caller's own user right after OlmMachine init.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device_keys]
              properties:
                device_keys:
                  type: object
                  description: 'Map of user_id -> array of device_ids, or an empty array meaning "all devices for that user".'
                  additionalProperties:
                    type: array
                    items:
                      type: string
      responses:
        '200':
          description: Device keys for the requested users
          content:
            application/json:
              schema:
                type: object
                properties:
                  device_keys:
                    type: object
                    description: 'user_id -> device_id -> opaque device_keys blob'
                  master_keys:
                    type: object
                    description: 'user_id -> opaque master cross-signing key blob, for any user that has bootstrapped cross-signing.'
                  self_signing_keys:
                    type: object
                    description: 'user_id -> opaque self-signing cross-signing key blob.'
                  user_signing_keys:
                    type: object
                    description: >
                      user_id -> opaque user-signing cross-signing key blob —
                      only populated for the caller's own user_id, matching
                      real Matrix's convention (the key that signs other
                      people's identities isn't broadcast to everyone who
                      queries you).
                  failures:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /keys/signing-keys/upload:
    post:
      tags: [keys]
      summary: Upload the caller's cross-signing public key triplet
      description: >
        Matches matrix-sdk-crypto-wasm's UploadSigningKeysRequest exactly —
        step 2 of 3 in OlmMachine.bootstrapCrossSigning() (step 1 is
        POST /keys/upload if a fresh device_keys upload is also pending; step
        3 is POST /keys/signatures/upload below). Opaque pass-through, same
        pattern as /keys/upload's device_keys — see
        docs/crypto-integration-notes.md's cross-signing grounding section.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                master_key:
                  type: object
                self_signing_key:
                  type: object
                user_signing_key:
                  type: object
      responses:
        '204':
          description: Stored
        '401':
          $ref: '#/components/responses/Unauthorized'
  /keys/signatures/upload:
    post:
      tags: [keys]
      summary: Upload cross-signing signatures for a device or identity
      description: >
        Matches matrix-sdk-crypto-wasm's SignatureUploadRequest — step 3 of
        bootstrapCrossSigning(), and also used ad hoc whenever a device or
        contact identity gets signed later (device linking, cross-signing a
        contact after SAS verification succeeds). Unlike every other
        endpoint in this API, this one can't be a pure opaque pass-through:
        the uploaded object only carries the *new* signature, so the server
        merges it into whatever's already stored (unioning `signatures[user_id]`
        entries) rather than overwriting — a live grounding script confirmed
        a naive overwrite drops a device's own pre-existing self-signature
        and silently corrupts the record for every future viewer. See
        docs/crypto-integration-notes.md's cross-signing grounding section
        (point 2). The target user_id is often NOT the caller — e.g. Bob
        uploading a signature over Alice's master key after verifying her is
        the normal, expected shape of this call, not a permissions bug; the
        server never validates signature cryptographic correctness itself
        (same "dumb pipe" philosophy as elsewhere), a bad signature just
        fails client-side verification for whoever fetches it back.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >
                user_id -> (device_id | cross-signing key's own public key
                value) -> the signed key object with the new signature
                included.
              additionalProperties:
                type: object
                additionalProperties:
                  type: object
      responses:
        '200':
          description: Signatures merged (per-target failures reported, not a hard error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  failures:
                    type: object
                    description: 'user_id -> key_id -> error, for any target that could not be matched to a known device or master key.'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /keys/claim:
    post:
      tags: [keys]
      summary: Claim one one-time key per requested device, to start new sessions
      description: >
        Matches matrix-sdk-crypto-wasm's KeysClaimRequest exactly. Claim is
        atomic and race-safe (verified with a real concurrent-claim test
        during implementation of the original /prekeys endpoint this
        replaces) — the same one-time key is never handed out twice.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [one_time_keys]
              properties:
                one_time_keys:
                  type: object
                  description: 'user_id -> device_id -> algorithm (always "signed_curve25519" for this stack)'
                timeout:
                  type: integer
      responses:
        '200':
          description: Claimed keys, one per device that had one available
          content:
            application/json:
              schema:
                type: object
                properties:
                  one_time_keys:
                    type: object
                    description: 'user_id -> device_id -> {keyId: opaque key object}. A device with no keys left is simply omitted, not an error.'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Device-scoped session token, issued by /register or /devices/link.
        Implemented as an opaque random token mapped to a device ID in Redis
        (backend/src/lib/auth.ts) — no expiry or refresh yet, that's still an
        open design question, not just an unimplemented one.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid session token
...